+++ lib/runners/cli.rb
@@ -63,7 +63,7 @@ module Runners
 
         result.tap do
           finished_at = Time.now
-          trace_writer.header "Analysis finished", recorded_at: finished_at
+          trace_writer.header "Analysis finished"
 
           if result.is_a? Results::Success
             trace_writer.message "#{result.issues.size} issue(s) found."
@@ -71,6 +71,7 @@ module Runners
 
           trace_writer.message "Finished at #{finished_at.utc}"
           trace_writer.message "Elapsed time: #{format_duration(finished_at - started_at)}"
+          trace_writer.finish started_at: started_at, finished_at: finished_at
         end
       end
     ensure