+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - **ESLint** Add `eslint-config-react-app` as pre-installed [#1902](https://github.com/sider/runners/pull/1902)
 - Fix checking unsupported tools [#1911](https://github.com/sider/runners/pull/1911)
 - Add `Runners::Config#exclude_branch?` method [#1913](https://github.com/sider/runners/pull/1913)
+- Add a new trace schema `finish` [#1914](https://github.com/sider/runners/pull/1914)
 
 ## 0.40.1